requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
environs==14.1.1
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
fastapi_cors==0.0.6
filelock==3.17.0
fsspec==2025.2.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.62.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
praw==7.8.1
prawcore==2.4.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
tweepy==4.15.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
update-checker==0.18.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0
